requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
langchain==0.2.0
langchain-community==0.2.0
sentence-transformers==2.7.0
weaviate-client==4.5.4
pypdf==4.2.0
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.3.4
structlog==24.1.0
prometheus_fastapi_instrumentator==6.1.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp==1.25.0
cohere==5.3.4
httpx==0.27.0
groq==0.4.1

# Rate limiting and reliability
tenacity==8.2.3

# Document processing
python-docx==1.1.2
python-pptx==0.6.23

# Article extraction and web scraping
beautifulsoup4==4.12.3
lxml==5.3.0
requests==2.32.3
